action-creators and constants broken up into separate files/folder
