+++ app/action-creators/category.jsx
@@ -1,0 +1,12 @@
+import {GET_CATEGORIES} from '../constants.jsx';
+
+export const getCategories = () =>
+  dispatch =>
+    axios.get('api/categories') //trying to destructure to access data
+      .then(({data}) => dispatch(listCategories(data)))
+      .catch((err) =>console.log(err) )
+
+export const listCategories = categoryList => ({
+  type: GET_CATEGORIES,
+  list: categoryList,
+})
