+++ app/constants.jsx
@@ -1,0 +1,1 @@
+export const GET_CATEGORIES = 'GET_CATEGORIES';
